gas md_apply_fix value casts
These are all innocuous but unneeded.  pdp11 and ppc are only formatting.

+++ gas/config/tc-metag.c
@@ -7029,7 +7029,7 @@ void
 md_apply_fix (fixS *fixP, valueT *valP, segT seg ATTRIBUTE_UNUSED)
 {
   char *buf = fixP->fx_where + fixP->fx_frag->fr_literal;
-  int value = (int)*valP;
+  int value = *valP;
 
   switch (fixP->fx_r_type)
     {
